Add Encoding property to WebSocketsServer
Add an Encoding property to the WebSocketsServer class, which inherited classes can use to decode the incoming byte[] to string in OnMessageReceived and OnFrameReceived.

The justification is that without this property, every implementation which deals with strings being received needs to define their own UTF8 encoding, while the WebSocket standard specifies that UTF8 must (should?) be used. It's convenient to have this encoding ready in the base class, so the programmer doesn't have to know or look up which encoding is used.

This also adds a protected setter just in case some library client wants to be dealing with non-UTF8 encoded strings and wants to use the Send method. The main justification is not defining UTF8 encoding twice, so imo the setter can be removed until someone has a valid use case for that.